add form error tracker + form trackers refactoring
