+++ src/analytics/formTracker/formTracker.test.ts
@@ -1,28 +1,12 @@
 import { describe, expect, jest, test } from "@jest/globals";
-import { FormResponseTracker } from "./formTracker";
+import { FormTracker } from "./formTracker";
 import { FormEventInterface, FormField } from "./formTracker.interface";
 
 window.DI = { analyticsGa4: { cookie: { consent: true } } };
 
-describe("FormResponseTracker", () => {
-  const action = new Event("submit", {
-    bubbles: true,
-    cancelable: true,
-  });
-
-  const spy = jest.spyOn(FormResponseTracker.prototype, "pushToDataLayer");
-  const constructorSpy = jest.spyOn(
-    FormResponseTracker.prototype,
-    "initialiseEventListener",
-  );
-
-  test("new instance should call initialiseEventListener", () => {
-    const instance = new FormResponseTracker();
-    expect(instance.initialiseEventListener).toBeCalled();
-  });
-
+describe("FormTracker", () => {
   test("getFields should return a list of fields objects", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const form = document.createElement("form");
     form.innerHTML =
       '<input id="test" name="test" value="test value" type="text"/>';
@@ -38,7 +22,7 @@ describe("FormResponseTracker", () => {
   });
 
   test("getFieldValue should return field value", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const fields: FormField[] = [
       { id: "test", name: "test", value: "test value", type: "test" },
     ];
@@ -46,7 +30,7 @@ describe("FormResponseTracker", () => {
   });
 
   test("getFieldType should return free text field if type is text", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const fields: FormField[] = [
       { id: "test", name: "test", value: "test value", type: "text" },
     ];
@@ -54,7 +38,7 @@ describe("FormResponseTracker", () => {
   });
 
   test("getFieldType should return free text field if type is textarea", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const fields: FormField[] = [
       { id: "test", name: "test", value: "test value", type: "textarea" },
     ];
@@ -62,7 +46,7 @@ describe("FormResponseTracker", () => {
   });
 
   test("getFieldType should return drop-down list if type is select-one", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const fields: FormField[] = [
       { id: "test", name: "test", value: "test value", type: "select-one" },
     ];
@@ -70,7 +54,7 @@ describe("FormResponseTracker", () => {
   });
 
   test("getFieldType should return checkbox if type is checkbox", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const fields: FormField[] = [
       { id: "test", name: "test", value: "test value", type: "checkbox" },
     ];
@@ -78,7 +62,7 @@ describe("FormResponseTracker", () => {
   });
 
   test("getFieldType should return radio if type is radio", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const fields: FormField[] = [
       { id: "test", name: "test", value: "test value", type: "radio" },
     ];
@@ -86,192 +70,11 @@ describe("FormResponseTracker", () => {
   });
 
   test("getFieldLabel should return field label", () => {
-    const instance = new FormResponseTracker();
+    const instance = new FormTracker();
     const label = document.createElement("label");
     label.innerHTML = "test label";
     label.textContent = "test label";
     document.body.appendChild(label);
     expect(instance.getFieldLabel()).toBe("test label");
-  });
-});
-describe("form with input checkbox", () => {
-  const action = new Event("submit", {
-    bubbles: true,
-    cancelable: true,
-  });
-
-  const spy = jest.spyOn(FormResponseTracker.prototype, "pushToDataLayer");
-
-  test("datalayer event should be defined", () => {
-    const instance = new FormResponseTracker();
-    document.body.innerHTML = "";
-    document.body.innerHTML =
-      '<form action="/test-url" method="post">' +
-      "  <legend>test label questions</legend>" +
-      '  <label for="question-1">test label question 1</label>' +
-      '  <input type="checkbox" id="question-1" name="question-1" value="test value" checked/>' +
-      '  <label for="question-2">test label question 2</label>' +
-      '  <input type="checkbox" id="question-2" name="question-2" value="test value"/>' +
-      '  <button id="button" type="submit">submit</button>' +
-      "</form>";
-    document.dispatchEvent(action);
-
-    const dataLayerEvent: FormEventInterface = {
-      event: "event_data",
-      event_data: {
-        event_name: "form_response",
-        type: "checkbox",
-        url: "undefined",
-        text: "test label question 1",
-        section: "test label questions",
-        action: "undefined",
-        external: "undefined",
-      },
-    };
-    expect(instance.pushToDataLayer).toBeCalledWith(dataLayerEvent);
-  });
-});
-describe("form with input text", () => {
-  const action = new Event("submit", {
-    bubbles: true,
-    cancelable: true,
-  });
-
-  const spy = jest.spyOn(FormResponseTracker.prototype, "pushToDataLayer");
-
-  test("datalayer event should be defined", () => {
-    const instance = new FormResponseTracker();
-    document.body.innerHTML =
-      '<form action="/test-url" method="post">' +
-      '  <label for="username">test label username</label>' +
-      '  <input type="text" id="username" name="username" value="test value"/>' +
-      '  <button id="button" type="submit">submit</button>' +
-      "</form>";
-    document.dispatchEvent(action);
-
-    const dataLayerEvent: FormEventInterface = {
-      event: "event_data",
-      event_data: {
-        event_name: "form_response",
-        type: instance.FREE_TEXT_FIELD_TYPE,
-        url: "undefined",
-        text: "test value",
-        section: "test label username",
-        action: "undefined",
-        external: "undefined",
-      },
-    };
-    expect(instance.pushToDataLayer).toBeCalledWith(dataLayerEvent);
-  });
-});
-
-describe("test disable free text tracking option", () => {
-  const action = new Event("submit", {
-    bubbles: true,
-    cancelable: true,
-  });
-
-  const spy = jest.spyOn(FormResponseTracker.prototype, "trackFormResponse");
-
-  test("pushToDataLayer should not be called with free text value", () => {
-    const instance = new FormResponseTracker({ disableFreeTextTracking: true });
-    document.body.innerHTML =
-      '<form action="/test-url" method="post">' +
-      '  <label for="username">test label username</label>' +
-      '  <input type="text" id="username" name="username" value="test no value"/>' +
-      '  <button id="button" type="submit">submit</button>' +
-      "</form>";
-    document.dispatchEvent(action);
-    expect(instance.trackFormResponse).toReturnWith(false);
-  });
-});
-
-describe("form with input textarea", () => {
-  const action = new Event("submit", {
-    bubbles: true,
-    cancelable: true,
-  });
-
-  const spy = jest.spyOn(FormResponseTracker.prototype, "pushToDataLayer");
-
-  test("datalayer event should be defined", () => {
-    const instance = new FormResponseTracker();
-    document.body.innerHTML =
-      '<form action="/test-url" method="post">' +
-      '  <label for="username">test label username</label>' +
-      '  <textarea id="username" name="username" value="test value"/>test value</textarea>' +
-      '  <button id="button" type="submit">submit</button>' +
-      "</form>";
-    document.dispatchEvent(action);
-
-    const dataLayerEvent: FormEventInterface = {
-      event: "event_data",
-      event_data: {
-        event_name: "form_response",
-        type: instance.FREE_TEXT_FIELD_TYPE,
-        url: "undefined",
-        text: "test value",
-        section: "test label username",
-        action: "undefined",
-        external: "undefined",
-      },
-    };
-    expect(instance.pushToDataLayer).toBeCalledWith(dataLayerEvent);
-  });
-});
-
-describe("form with dropdown", () => {
-  const action = new Event("submit", {
-    bubbles: true,
-    cancelable: true,
-  });
-
-  const spy = jest.spyOn(FormResponseTracker.prototype, "pushToDataLayer");
-
-  test("datalayer event should be defined", () => {
-    const instance = new FormResponseTracker();
-    document.body.innerHTML =
-      '<form action="/test-url" method="post">' +
-      '  <label for="username">test label username</label>' +
-      '  <select id="username" name="username"><option value="test value">test value</option><option value="test value2" selected>test value2</option></select>' +
-      '  <button id="button" type="submit">submit</button>' +
-      "</form>";
-    document.dispatchEvent(action);
-
-    const dataLayerEvent: FormEventInterface = {
-      event: "event_data",
-      event_data: {
-        event_name: "form_response",
-        type: "drop-down list",
-        url: "undefined",
-        text: "test value2",
-        section: "test label username",
-        action: "undefined",
-        external: "undefined",
-      },
-    };
-    expect(instance.pushToDataLayer).toBeCalledWith(dataLayerEvent);
-  });
-});
-
-describe("Cookie Management", () => {
-  const action = new Event("submit", {
-    bubbles: true,
-    cancelable: true,
-  });
-  const spy = jest.spyOn(FormResponseTracker.prototype, "trackFormResponse");
-  const instance = new FormResponseTracker();
-
-  test("trackFormResponse should return false if not cookie consent", () => {
-    window.DI.analyticsGa4.cookie.consent = false;
-    document.body.innerHTML =
-      '<form action="/test-url" method="post">' +
-      '  <label for="username">test label username</label>' +
-      '  <select id="username" name="username"><option value="test value">test value</option><option value="test value2" selected>test value2</option></select>' +
-      '  <button id="button" type="submit">submit</button>' +
-      "</form>";
-    document.dispatchEvent(action);
-
-    expect(instance.trackFormResponse).toReturnWith(false);
   });
 });
